Got stuff working on windows
So building on windows involves first getting mingw of some sort, installing a
version of make that works mostly like it does on unix (in mingw32, this is the
msys one, NOT the one that comes with it in mingw32/bin!), installing
development headers for SDL2 using said make, and finally installing git and
getting this repository and running make to compile the stuff. Note that it's
important that this happens in that order - we don't want to be using any of the
stuff that comes with git for windows - let that be self-contained.

I've successfully compiled the project so far for windows and ran it, but
noticed some differences from the way it behaves on linux. For starters, there's
no frame whatsoever to begin with (or if there is, the frame is going off the
top of the screen), so consequently the window can't be moved
currently. Additionally, on clicking the window and pushing buttons on the
keyboard, we don't get any output about getting events. At least, that's what it
seems like until after enough events about 300 lines of output get sent all at
once. I attempted to fix this by flushing stdout after each event is processed,
but it didn't seem to do anything.